+++ CHANGELOG.md
@@ -59,6 +59,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([`#316`](https://github.com/polybar/polybar/issues/316))
 - Added .ini extension check to the default config search.
   ([`#2323`](https://github.com/polybar/polybar/issues/2323))
+- IPC commands to change visibility of modules
+  (`hide.<name>`, `show.<name>`, and `toggle.<name>`)
+  ([`#2108`](https://github.com/polybar/polybar/issues/2108))
 
 ### Changed
 - Slight changes to the value ranges the different ramp levels are responsible